requirements.txt
anthropic>=0.40.0
datasets>=2.19.0
google-genai>=1.0.0
numpy>=1.26.0
openai>=1.0.0
pyyaml>=6.0.1
scipy>=1.13.0
